refactor(knowledge-store): retire the duplicate path_resolver module
Every path helper the agent runtime exposed already lives in
app.knowledge_store.paths, so path_resolver was a second, drifting copy of
the store's path law. Point all 28 importers at the package and delete the
425-line duplicate.

This also closes a layering inversion: index/{converge,rows,project} were
importing up into the agent runtime for path helpers that belong to their
own module.

The one behavioural difference was parse_documents_path, which strips the
.xml extension and " (<id>)" suffix to form a title — the contract the
indexer and its tests depend on. It moves to paths/legacy.py (it is xml
aware, so legacy) and becomes the exported one; store_path's unused raw
splitter is dropped.

virtual_path_to_doc is now the healed, column-first resolver everywhere,
a strict superset of the old hash/suffix/title lookup.
