+++ surfsense_backend/app/agents/chat/multi_agent_chat/subagents/shared/run_reader.py
@@ -166,7 +166,7 @@ async def _save_export_document(
     from app.agents.chat.multi_agent_chat.main_agent.middleware.kb_persistence.middleware import (
         _create_document,
     )
-    from app.agents.chat.runtime.path_resolver import DOCUMENTS_ROOT
+    from app.knowledge_store.paths import DOCUMENTS_ROOT
     from app.db import async_session_maker
 
     path = virtual_path.strip()
