+++ surfsense_backend/app/agents/chat/runtime/references/folders.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.agents.chat.runtime.path_resolver import DOCUMENTS_ROOT, PathIndex
+from app.knowledge_store.paths import DOCUMENTS_ROOT, PathIndex
 from app.db import Folder
 
 from .models import FolderReference
