+++ surfsense_backend/app/services/revert_service.py
@@ -38,7 +38,7 @@ from typing import Any, Literal
 from sqlalchemy import delete, select
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.agents.chat.runtime.path_resolver import (
+from app.knowledge_store.paths import (
     DOCUMENTS_ROOT,
     safe_filename,
     safe_folder_segment,
